WIP: consolidated CVR class into Audit class
